fixed issue with duplicated exports and removal of the default ones - exported functions are now all passed in an array - exports array defaults to DllRegisterServer, DllGetClassObject, DllUnregisterServer - ExportsFromFile is not in utils.go
